+++ packages/sensors_plus_platform_interface/CHANGELOG.md
@@ -1,3 +1,7 @@
+## 0.2.0
+
+- Renamed Method Channel
+
 ## 0.1.0
 
 - Transfer to plus-plugins monorepo